win32 support module updated ...
git-svn-id: https://svn.code.sf.net/p/freedos-32/code/trunk@701 85e9c19b-ca11-0410-bd92-afdf7feba87b

+++ fd32/modules/winb/winb.h
@@ -26,14 +26,14 @@ struct symbol {
  */
 typedef struct process_info {
   struct process_info *prev;
-  DWORD type;
-  void *psp;      /* Optional DOS PSP */
-  void *cds_list; /* Under DOS this is a global array            */
+  uint32_t type;
+  void *psp;		/* Optional DOS PSP */
+  void *cds_list;	/* Under DOS this is a global array */
   char *args;
   char *filename;
-  DWORD memlimit;
+  uint32_t memlimit;
   void *jft;
-  WORD  jft_size;
+  uint32_t jft_size;
 } process_info_t;
 
 process_info_t *fd32_get_current_pi(void);
